Add MIT License to project
A new LICENSE file has been added, specifying the MIT License for the project. This license grants permission to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software. It also includes a disclaimer of warranties and liabilities. The copyright is attributed to Nominal Solutions LLC for the year 2025.